+++ src/RobotInterface/include/BipedalLocomotion/RobotInterface/ICameraBridge.h
@@ -71,7 +71,7 @@ public:
      * Initialize estimator
      * @param[in] handler Parameters handler
      */
-    virtual bool initialize(std::weak_ptr<BipedalLocomotion::ParametersHandler::IParametersHandler> handler) = 0;
+    virtual bool initialize(std::weak_ptr<const BipedalLocomotion::ParametersHandler::IParametersHandler> handler) = 0;
 
     /**
      * Get rgb cameras
@@ -160,6 +160,11 @@ protected:
     virtual bool populateCameraBridgeMetaDataFromConfig(std::weak_ptr<BipedalLocomotion::ParametersHandler::IParametersHandler> handler,
                                                         CameraBridgeMetaData& cameraBridgeMetaData) { return true; };
 
+    /**
+     * Get the stored metadata.
+     * @return a const reference to the metadata
+     */
+    virtual const CameraBridgeMetaData& getMetaData() const = 0;
 
 };
 } // namespace RobotInterface