+++ src/RobotInterface/include/BipedalLocomotion/RobotInterface/IPointCloudBridge.h
@@ -6,12 +6,12 @@
 #ifndef BIPEDAL_LOCOMOTION_ROBOT_INTERFACE_IPCL_BRIDGE_H
 #define BIPEDAL_LOCOMOTION_ROBOT_INTERFACE_IPCL_BRIDGE_H
 
-#include <memory>
-#include <string>
-#include <vector>
-#include <unordered_map>
-#include <optional>
 #include <functional>
+#include <memory>
+#include <optional>
+#include <string>
+#include <unordered_map>
+#include <vector>
 
 #include <BipedalLocomotion/ParametersHandler/IParametersHandler.h>
 #include <pcl/point_cloud.h>
@@ -30,14 +30,13 @@ struct PCLDeviceLists
     std::vector<std::string> pclDevList; /**< list of PCL cameras attached to the bridge */
 };
 
-
 /**
  * Meta data struct to hold list of sensors and configured options
  * available from the Sensor bridge interface
  */
 struct PCLBridgeMetaData
 {
-  PCLDeviceLists pclDevList;
+    PCLDeviceLists pclDevList;
 };
 
 /**
@@ -56,28 +55,34 @@ public:
      * Initialize estimator
      * @param[in] handler Parameters handler
      */
-    virtual bool initialize(std::weak_ptr<BipedalLocomotion::ParametersHandler::IParametersHandler> handler) = 0;
+    virtual bool
+    initialize(std::weak_ptr<const BipedalLocomotion::ParametersHandler::IParametersHandler> handler)
+        = 0;
 
     /**
      * Get RGBD cameras
      * @param[out] rgbdCamerasList list of rgbd cameras attached to the bridge
      * @return  true/false in case of success/failure
      */
-    virtual bool getPCLDevicesList(std::vector<std::string>& pclDevList) { return false; };
+    virtual bool getPCLDevicesList(std::vector<std::string>& pclDevList)
+    {
+        return false;
+    };
 
-
-    virtual bool getPointCloud(const std::string& pclDev,
-                               pcl::PointCloud<pcl::PointXYZRGB>::Ptr coloredPointCloud,
-                               std::optional<std::reference_wrapper<double>> receiveTimeInSeconds = {}) { return false; };
+    virtual bool
+    getPointCloud(const std::string& pclDev,
+                  pcl::PointCloud<pcl::PointXYZRGB>::Ptr coloredPointCloud,
+                  std::optional<std::reference_wrapper<double>> receiveTimeInSeconds = {})
+    {
+        return false;
+    };
 
     /**
      * Destructor
      */
     virtual ~IPointCloudBridge() = default;
-
 };
 } // namespace RobotInterface
 } // namespace BipedalLocomotion
 
 #endif // BIPEDAL_LOCOMOTION_ROBOT_INTERFACE_IPCL_BRIDGE_H
-